requirements.txt
flask
google-generativeai
python-telegram-bot==20.3
requests